put new routes in the lib folder.  removed console log's and cleaned the code.
