+++ lib/createMemo.js
@@ -3,8 +3,6 @@
 const client = require('./Client');
 
 function createMemo(request, response) {
-  console.log('request.body: ', request.body);
-
   var originOfRequest = request.headers.referer;
   var originOfRequestRefactor = originOfRequest.split('/');
   var originLastSection = originOfRequestRefactor[4];
@@ -14,10 +12,6 @@ function createMemo(request, response) {
   const memo = request.body.memo.toLowerCase();
   const choreType = request.body.choreType.toLowerCase();
   const id = userId;
-
-  console.log('memo: ', memo);
-  console.log('choreType: ', choreType);
-  console.log('id: ', id);
 
   const VALUES = [id, memo, choreType];
   const SQL = 'INSERT INTO memos(id, memo, chore_type) VALUES($1, $2, $3) RETURNING id, memo, chore_type, memo_id;';
