+++ lib/getGroceries.js
@@ -4,7 +4,7 @@ const client = require('./Client');
 
 function getGroceries(request, response) {
   var idInput = request.params.idInput;
-  const SQL = `SELECT * FROM memos INNER JOIN users ON users.id=memos.id WHERE users.id=${idInput};`;
+  const SQL = `SELECT * FROM memos INNER JOIN users ON users.id=memos.id WHERE users.id=${idInput} AND chore_type=\'groceries\';`;
 
   client.query(SQL)
     .then(results => {
