+++ lib/getMemos.js
@@ -3,10 +3,8 @@
 const client = require('./Client');
 
 function getMemos(request, response) {
-
   const idInput = request.params.idInput;
-  console.log('line 13 - idInput: ', idInput);
-  const SQL = `SELECT * FROM memos INNER JOIN users ON users.id=memos.id WHERE users.id=${idInput};`;
+  const SQL = `SELECT * FROM memos INNER JOIN users ON users.id=memos.id WHERE users.id=${idInput} ORDER BY chore_type DESC;`;
 
   client.query(SQL)
     .then(results => {
