+++ views/pages/dropDownPage.ejs
@@ -1,11 +1,11 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="/css/pages/dropDownPage.css">
-  <title>Document</title>
-</head>
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/css/pages/dropDownPage.css">
+    <title>Document</title>
+  </head>
   <body>
     <form action="/groceries/<%= userId %>" method="get">
       <button id="groceriesButton" name="groceriesButton">Groceries</button>
